requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Werkzeug==2.3.6
zipp==3.16.2
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
psycopg2==2.9.6
gunicorn==21.2.0
marshmallow==3.21.2
marshmallow-enum==1.5.1
requests==2.32.2
flask_cors==4.0.1
cryptography==3.4.8
redis==5.0.4